+++ library/src/main/java/ch/decent/sdk/net/model/request/GetNftDataCount.kt
@@ -1,11 +1,10 @@
 package ch.decent.sdk.net.model.request
 
 import ch.decent.sdk.net.model.ApiGroup
-import java.math.BigInteger
 
-//@UInt64
-internal object GetNftDataCount : BaseRequest<BigInteger>(
+//@UInt64, capped by max instance id so Long is safe
+internal object GetNftDataCount : BaseRequest<Long>(
     ApiGroup.DATABASE,
     "get_non_fungible_token_data_count",
-    BigInteger::class.java
+    Long::class.java
 )
